src/Backend/GTFS_RT/requirements.txt
gtfs-realtime-bindings>=1.0
requests>=2.28
psycopg2-binary>=2.9
python-dotenv>=1.0
